# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Budgie",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/budgets": {
      "get": {
        "tags": [
          "budgets"
        ],
        "summary": "List the monthly budgets",
        "operationId": "list_budgets_api_budgets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Budget"
                  },
                  "type": "array",
                  "title": "Response List Budgets Api Budgets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "budgets"
        ],
        "summary": "Replace the whole budget list",
        "operationId": "put_budgets_api_budgets_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/Budget"
                },
                "type": "array",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Budget"
                  },
                  "type": "array",
                  "title": "Response Put Budgets Api Budgets Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/goal": {
      "get": {
        "tags": [
          "budgets"
        ],
        "summary": "Get the savings goal",
        "operationId": "get_goal_api_goal_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "budgets"
        ],
        "summary": "Set the savings goal",
        "operationId": "put_goal_api_goal_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Goal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Budget": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "groceries.fresh",
              "groceries.staples",
              "snacks_sweets",
              "drinks",
              "alcohol",
              "tobacco",
              "household",
              "personal_care",
              "health",
              "eating_out",
              "transport",
              "clothing",
              "electronics",
              "other"
            ],
            "title": "Category"
          },
          "monthly_limit": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Monthly Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "monthly_limit"
        ],
        "title": "Budget",
        "description": "A monthly limit for one spending category. Sent and returned in `/budgets`."
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "bad_request",
              "not_found",
              "method_not_allowed",
              "invalid_state",
              "file_too_large",
              "unsupported_file",
              "uncategorized_items",
              "incomplete_expense",
              "validation_error",
              "storage_error",
              "internal_error",
              "not_implemented"
            ],
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorBody",
        "description": "The body of every non-2xx response (error-format.md)."
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "examples": [
              "line_items.0.amount"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "description": "One failed field of a `validation_error`."
      },
...
      "Goal": {
        "properties": {
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          },
          "target_date": {
            "type": "string",
            "format": "date",
            "title": "Target Date"
          },
          "monthly_income": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Income"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_amount",
          "target_date"
        ],
        "title": "Goal",
        "description": "The savings goal. Sent and returned in `/goal`; `monthly_income` is always dumped."
      },
...
    }
  }
}